# DDD - Falta Print das Telas/Documentação API/Raízes Solidárias.postman_collection.json-OpenApi3Json.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Raízes Solidárias",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "variables": {}
    },
    {
      "url": "https://localhost:8080",
      "variables": {}
    }
  ],
  "paths": {
...
    "/plantio_voluntario": {
      "get": {
        "tags": [
          "Plantio_Voluntario"
        ],
        "summary": "Listar Plantio_Voluntarios",
        "operationId": "ListarPlantio_Voluntarios",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:8080",
            "variables": {}
          }
        ]
      },
      "post": {
        "tags": [
          "Plantio_Voluntario"
        ],
        "summary": "Cadastrar Plantio_Voluntario",
        "operationId": "CadastrarPlantio_Voluntario",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CadastrarPlantio_VoluntarioRequest"
                  },
                  {
                    "example": {
                      "plantio": {
                        "id_plantio": 1
                      },
                      "voluntario": {
                        "id_usuario": 2
                      }
                    }
                  }
                ]
              },
              "example": {
                "plantio": {
                  "id_plantio": 1
                },
                "voluntario": {
                  "id_usuario": 2
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:8080",
            "variables": {}
          }
        ]
      }
    },
...
    "/plantio_voluntario/2-5/3-4": {
      "put": {
        "tags": [
          "Plantio_Voluntario"
        ],
        "summary": "Atualizar Plantio_Voluntario",
        "operationId": "AtualizarPlantio_Voluntario",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AtualizarPlantio_VoluntarioRequest"
                  },
                  {
                    "example": {
                      "id_plantio_novo": 3,
                      "id_plantio_antigo": 2,
                      "id_usuario_novo": 4,
                      "id_usuario_antigo": 5
                    }
                  }
                ]
              },
              "example": {
                "id_plantio_novo": 3,
                "id_plantio_antigo": 2,
                "id_usuario_novo": 4,
                "id_usuario_antigo": 5
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:8080",
            "variables": {}
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Voluntario": {
        "title": "Voluntario",
        "required": [
          "id_usuario"
        ],
        "type": "object",
        "properties": {
          "id_usuario": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "id_usuario": 17
        }
      },
...
      "Plantio": {
        "title": "Plantio",
        "required": [
          "id_plantio"
        ],
        "type": "object",
        "properties": {
          "id_plantio": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "id_plantio": 8
        }
      },
      "AtualizarPlantio_VoluntarioRequest": {
        "title": "AtualizarPlantio_VoluntarioRequest",
        "required": [
          "id_plantio_novo",
          "id_plantio_antigo",
          "id_usuario_novo",
          "id_usuario_antigo"
        ],
        "type": "object",
        "properties": {
          "id_plantio_novo": {
            "type": "integer",
            "format": "int32"
          },
          "id_plantio_antigo": {
            "type": "integer",
            "format": "int32"
          },
          "id_usuario_novo": {
            "type": "integer",
            "format": "int32"
          },
          "id_usuario_antigo": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "id_plantio_novo": 3,
          "id_plantio_antigo": 2,
          "id_usuario_novo": 4,
          "id_usuario_antigo": 5
        }
      },
      "CadastrarPlantio_VoluntarioRequest": {
        "title": "CadastrarPlantio_VoluntarioRequest",
        "required": [
          "plantio",
          "voluntario"
        ],
        "type": "object",
        "properties": {
          "plantio": {
            "$ref": "#/components/schemas/Plantio"
          },
          "voluntario": {
            "$ref": "#/components/schemas/Voluntario"
          }
        },
        "example": {
          "plantio": {
            "id_plantio": 1
          },
          "voluntario": {
            "id_usuario": 2
          }
        }
      },
...
    }
  },
...
}